misc changes * OggFromWebMWriter: rewrite (again), reduce iterations over the input. Works as-is (video streams are not supported) * WebMReader: use int for SimpleBlock.dataSize instead of long * Download Recovery: allow recovering uninitialized downloads * check range-requests using HEAD method instead of GET * DownloadRunnableFallback: add workaround for 32kB/s issue, unknown issue origin, wont fix * reporting downloads errors now include the source url with the selected quality and format

+++ app/src/main/java/org/schabi/newpipe/streams/WebMReader.java
@@ -368,7 +368,7 @@ public class WebMReader {
         obj.trackNumber = readEncodedNumber();
         obj.relativeTimeCode = stream.readShort();
         obj.flags = (byte) stream.read();
-        obj.dataSize = (ref.offset + ref.size) - stream.position();
+        obj.dataSize = (int) ((ref.offset + ref.size) - stream.position());
         obj.createdFromBlock = ref.type == ID_Block;
 
         // NOTE: lacing is not implemented, and will be mixed with the stream data
@@ -465,7 +465,7 @@ public class WebMReader {
         public short relativeTimeCode;
         public long absoluteTimeCodeNs;
         public byte flags;
-        public long dataSize;
+        public int dataSize;
         private final Element ref;
 
         public boolean isKeyframe() {
